Linux compatibility

+++ InternetHeaders/InternetTypes.h
@@ -3,7 +3,7 @@
 
 typedef unsigned char byte;
 typedef unsigned short word;
-typedef unsigned long dword;
+typedef unsigned int dword;
 typedef unsigned long long qword;
 
 #endif
